Command Logic
Working version of command logic to fetch the most starred PHP repos and store them into the database using the PHPStars model.
Set the command to run so information is always current.
Created an invokable controller to easily pass data to the view
Set up migrations table to accept the data
Working on creating the views